requirements.txt
torch>=2.3.0
torchvision
torchaudio
transformers>=4.40.0
sentence-transformers
accelerate
fastapi
uvicorn
python-dotenv
pydantic
numpy
tqdm
requests
PyYAML
uvloop